just a little refactoring
